Various improvements and fixes.

+++ src/util.h
@@ -1,11 +1,11 @@
-float capRange(float min, float val, float max) {
+double capRange(double min, double val, double max) {
   if (val < min) return min;
   if (val > max) return max;
   return val;
 }
 
-float normaliseValueToRange(float min, float val, float max) {
-  float range = max - min;
+double normaliseValueToRange(double min, double val, double max) {
+  double range = max - min;
   while (val > max) val -= range;
   while (val < min) val += range;
   return val;
